Upgrade to rustc 1.16.0-dev (7bfe5c028 2017-01-23).

+++ src/tut03/glscene.rs
@@ -80,7 +80,7 @@ impl GLScene {
 
             // Send vertices to buffer.
             gl::BufferData(gl::ARRAY_BUFFER
-                , std::mem::size_of_val(&G_VERTEX_BUFFER_DATA) as i64
+                , std::mem::size_of_val(&G_VERTEX_BUFFER_DATA) as isize
                 , std::mem::transmute(&G_VERTEX_BUFFER_DATA)
                 , gl::STATIC_DRAW);
         }
